+++ racetrack_client/racetrack_client/main.py
@@ -1,17 +1,19 @@
+import sys
 from typing import List
 
 import typer
 
 from racetrack_client import __version__
+from racetrack_client.client.call import call_job
 from racetrack_client.client.deploy import BuildContextMethod, send_deploy_request, DeploymentError
-from racetrack_client.client.manage import JobTableColumn, move_job, delete_job, list_jobs
+from racetrack_client.client.manage import JobTableColumn, move_job, delete_job, list_jobs, complete_job_name
 from racetrack_client.client.logs import show_runtime_logs, show_build_logs
+from racetrack_client.client.run import run_job_locally
 from racetrack_client.client_config.auth import login_user_auth, logout_user_auth
 from racetrack_client.client_config.io import load_client_config
 from racetrack_client.client_config.update import set_credentials, set_current_remote, get_current_remote, set_config_url_alias
 from racetrack_client.plugin.bundler.bundle import bundle_plugin
 from racetrack_client.plugin.install import install_plugin, list_available_job_types, list_installed_plugins, uninstall_plugin
-from racetrack_client.client.run import run_job_locally
 from racetrack_client.log.exception import log_exception
 from racetrack_client.log.logs import configure_logs
 from racetrack_client.log.logs import get_logger
@@ -27,6 +29,7 @@ cli = typer.Typer(
     help='CLI client tool for managing workloads in Racetrack',
 )
 
+
 def main():
     try:
         cli()
@@ -40,7 +43,10 @@ def main():
 def _startup(
     verbose: bool = typer.Option(False, '-v', '--verbose', help='enable verbose mode'),
 ):
-    configure_logs(log_level='debug' if verbose else 'info')
+    if not sys.stdout.isatty():
+        configure_logs(log_level='error')
+    else:
+        configure_logs(log_level='debug' if verbose else 'info')
 
 
 @cli.command('deploy')
@@ -65,7 +71,7 @@ def _validate(
 
 @cli.command('logs', no_args_is_help=True)
 def _logs(
-    name: str = typer.Argument(..., show_default=False, help='name of the job'),
+    name: str = typer.Argument(..., show_default=False, help='name of the job', autocompletion=complete_job_name),
     version: str = typer.Option('latest', show_default=True, help='version of the job'),
     remote: str = typer.Option(default=None, show_default=False, help="Racetrack server's URL or alias name"),
     tail: int = typer.Option(20, '--tail', help='number of recent lines to show'),
@@ -77,7 +83,7 @@ def _logs(
 
 @cli.command('build-logs', no_args_is_help=True)
 def _build_logs(
-    name: str = typer.Argument(..., show_default=False, help='name of the job'),
+    name: str = typer.Argument(..., show_default=False, help='name of the job', autocompletion=complete_job_name),
     version: str = typer.Option('latest', show_default=True, help='version of the job'),
     remote: str = typer.Option(default=None, show_default=False, help="Racetrack server's URL or alias name"),
     tail: int = typer.Option(0, '--tail', help='number of recent lines to show, all logs by default'),
@@ -97,7 +103,7 @@ def _list_jobs(
 
 @cli.command('delete', no_args_is_help=True)
 def _delete_job(
-    name: str = typer.Argument(..., show_default=False, help='name of the job'),
+    name: str = typer.Argument(..., show_default=False, help='name of the job', autocompletion=complete_job_name),
     version: str = typer.Option(..., show_default=False, help='version of the job to delete'),
     remote: str = typer.Option(default=None, show_default=False, help="Racetrack server's URL or alias name"),
 ):
@@ -107,7 +113,7 @@ def _delete_job(
 
 @cli.command('move', no_args_is_help=True)
 def _move_job(
-    name: str = typer.Argument(..., show_default=False, help='name of the job'),
+    name: str = typer.Argument(..., show_default=False, help='name of the job', autocompletion=complete_job_name),
     version: str = typer.Option(..., show_default=False, help='version of the job to move out'),
     infrastructure: str = typer.Option(..., show_default=False, help='infrastructure target to move to'),
     remote: str = typer.Option(default=None, show_default=False, help="Racetrack server's URL or alias name"),
@@ -154,6 +160,7 @@ def _logout(
 cli_set = typer.Typer(no_args_is_help=True, help='Set global options of the Racetrack client')
 cli.add_typer(cli_set, name="set")
 
+
 @cli_set.command('remote', no_args_is_help=True)
 def _set_remote(
     remote: str = typer.Argument(..., show_default=False, help="Racetrack server's URL or alias name"),
@@ -185,6 +192,7 @@ def _set_alias(
 cli_get = typer.Typer(no_args_is_help=True, help='Read global options of the Racetrack client')
 cli.add_typer(cli_get, name="get")
 
+
 @cli_get.command('remote')
 def _get_remote():
     """Get current Racetrack's remote address"""
@@ -201,6 +209,7 @@ def _get_config():
 # racetrack plugin ...
 cli_plugin = typer.Typer(no_args_is_help=True, help='Manage Racetrack plugins')
 cli.add_typer(cli_plugin, name="plugin")
+
 
 @cli_plugin.command('install', no_args_is_help=True)
 def _install_plugin(
@@ -241,3 +250,16 @@ def _plugin_bundle(
 ):
     """Turn local plugin code into ZIP file"""
     bundle_plugin(workdir, out, plugin_version)
+
+
+@cli.command('call', no_args_is_help=True)
+def _call_job(
+    name: str = typer.Argument(..., show_default=False, help='name of the job', autocompletion=complete_job_name),
+    endpoint: str = typer.Argument(..., show_default=False, help='endpoint of the job to call, eg. /api/v1/perform'),
+    payload: str = typer.Argument(..., show_default=False, help='payload of the request in JSON or YAML format'),
+    version: str = typer.Option('latest', show_default=True, help='version of the job'),
+    remote: str = typer.Option(default=None, show_default=False, help="Racetrack server's URL or alias name"),
+    curl: bool = typer.Option(False, '--curl', help='generate a curl query instead of calling the job'),
+):
+    """Call an endpoint of a job"""
+    call_job(name, version, remote, endpoint, payload, curl)